requirements.txt
PyYAML
gnureadline